Refactor file descriptor handling and syscall return types
- Changed the type of file descriptor in the FILE structure from size_t to int for better compatibility with standard file descriptor conventions.
- Updated syscall functions to use int instead of ssize_t for file descriptor parameters and return types, ensuring consistency across the codebase.
- Modified related functions in the process and thread management to use pid_t instead of size_t for process and thread IDs.
- Adjusted syscall routines for file operations (open, read, write, etc.) to reflect the new int type for file descriptors.
- Enhanced error handling and return types in various syscall routines to improve clarity and maintainability.
- Updated function signatures in the virtual file system (VFS) layer to align with the new file descriptor type.

+++ module/src/libc/stdio.h
@@ -10,7 +10,7 @@
 /* Types */
 
 typedef struct {
-    size_t fd;
+    int fd;
 } FILE;
 
 typedef struct {
