+++ src/syscall/fs/entries.c
@@ -1,6 +1,6 @@
 #include "syscall/fs/syscall_fs.h"
 
-ssize_t syscall_routine_getdents(size_t fd, void *buf, size_t count, proc_t *proc, pml4_t *current_pml4) {
+ssize_t syscall_routine_getdents(int fd, void *buf, size_t count, proc_t *proc, pml4_t *current_pml4) {
     printf("Getdents syscall called on fd %p for process %s (pid %p) for %p bytes.\n", fd, proc->name, proc->pid, count);
 
     vfs_handle_t *handle = NULL;
@@ -38,12 +38,12 @@ ssize_t syscall_routine_getdents(size_t fd, void *buf, size_t count, proc_t *pro
     return res;
 }
 
-ssize_t syscall_routine_mkdir(const char *restrict path, mode_t mode, proc_t *proc) {
-    ssize_t res = 0;
+int syscall_routine_mkdir(const char *restrict path, mode_t mode, proc_t *proc) {
+    int res = 0;
 
     char path_abs[PATH_MAX] = { 0 };
 
-    res = (ssize_t)handle_relative_path(path, proc, path_abs);
+    res = handle_relative_path(path, proc, path_abs);
     if (res < 0) {
         printf("syscall_routine_mkdir: failed to handle relative path %s for process %s (pid %p), errno %d\n", path,
                proc->name, proc->pid, res);
@@ -53,6 +53,6 @@ ssize_t syscall_routine_mkdir(const char *restrict path, mode_t mode, proc_t *pr
     return vfs_mkdir(path_abs, proc->uid, 0 /* TODO: use egid here */, mode);
 }
 
-ssize_t syscall_routine_rmdir(const char *restrict path, proc_t *proc) {
+int syscall_routine_rmdir(const char *restrict path, proc_t *proc) {
     return -ENOSYS;
 }
