+++ src/syscall/fs/rw.c
@@ -11,12 +11,12 @@
 
 #include "plenjos/errno.h"
 
-ssize_t syscall_routine_read(size_t fd, void *buf, size_t count, proc_t *proc, pml4_t *current_pml4) {
+ssize_t syscall_routine_read(int fd, void *buf, size_t count, proc_t *proc, pml4_t *current_pml4) {
     printf("syscall_routine_read called with fd %p, buf %p, count %p\n", fd, buf, count);
 
     vfs_handle_t *handle = NULL;
 
-    ssize_t res = syscall_fs_helper_get_not_dir_handle(fd, proc, &handle);
+    ssize_t res = (ssize_t)syscall_fs_helper_get_not_dir_handle(fd, proc, &handle);
     if (res < 0) {
         printf("syscall_routine_read: failed to get non-directory handle\n");
         return res;
@@ -47,10 +47,10 @@ ssize_t syscall_routine_read(size_t fd, void *buf, size_t count, proc_t *proc, p
     return res;
 }
 
-ssize_t syscall_routine_write(size_t fd, const void *buf, size_t count, proc_t *proc) {
+ssize_t syscall_routine_write(int fd, const void *buf, size_t count, proc_t *proc) {
     vfs_handle_t *handle = NULL;
 
-    ssize_t res = syscall_fs_helper_get_not_dir_handle(fd, proc, &handle);
+    ssize_t res = (ssize_t)syscall_fs_helper_get_not_dir_handle(fd, proc, &handle);
     if (res < 0) {
         printf("syscall_routine_write: failed to get non-directory handle\n");
         return res;
@@ -59,10 +59,10 @@ ssize_t syscall_routine_write(size_t fd, const void *buf, size_t count, proc_t *
     return vfs_write(handle, buf, count);
 }
 
-ssize_t syscall_routine_lseek(size_t fd, ssize_t offset, int whence, proc_t *proc) {
+off_t syscall_routine_lseek(int fd, off_t offset, int whence, proc_t *proc) {
     vfs_handle_t *handle = NULL;
 
-    ssize_t res = syscall_fs_helper_get_handle_and_check_backing_node(fd, proc, &handle);
+    off_t res = (off_t)syscall_fs_helper_get_handle_and_check_backing_node(fd, proc, &handle);
     if (res < 0) {
         printf("syscall_routine_lseek: failed to get handle with backing node\n");
         return res;
