+++ src/syscall/fs/stat.c
@@ -10,8 +10,8 @@
 #include "vfs/vfs.h"
 
 // Handle must be valid and have a backing node
-static ssize_t syscall_routine_stat_from_handle(vfs_handle_t *handle, uint64_t out_stat_ptr, pml4_t *current_pml4) {
-    ssize_t res = 0;
+static int syscall_routine_stat_from_handle(vfs_handle_t *handle, uint64_t out_stat_ptr, pml4_t *current_pml4) {
+    int res = 0;
 
     struct kstat kstat = { 0 };
 
@@ -21,7 +21,7 @@ static ssize_t syscall_routine_stat_from_handle(vfs_handle_t *handle, uint64_t o
     kstat.uid  = node->uid;
     kstat.gid  = node->gid;
 
-    res = (ssize_t)copy_to_user_buf((void *)out_stat_ptr, &kstat, sizeof(struct kstat), current_pml4);
+    res = copy_to_user_buf((void *)out_stat_ptr, &kstat, sizeof(struct kstat), current_pml4);
     if (res < 0) {
         return res;
     }
@@ -29,12 +29,12 @@ static ssize_t syscall_routine_stat_from_handle(vfs_handle_t *handle, uint64_t o
     return res;
 }
 
-ssize_t syscall_routine_stat(const char *restrict path, uint64_t out_stat_ptr, proc_t *proc, pml4_t *current_pml4) {
-    ssize_t res = 0;
+int syscall_routine_stat(const char *restrict path, uint64_t out_stat_ptr, proc_t *proc, pml4_t *current_pml4) {
+    int res = 0;
 
     char path_abs[PATH_MAX] = { 0 };
 
-    res = (ssize_t)handle_relative_path(path, proc, path_abs);
+    res = handle_relative_path(path, proc, path_abs);
     if (res < 0) {
         printf("syscall_routine_stat: failed to handle relative path %s for process %s (pid %p), errno %d\n", path,
                proc->name, proc->pid, res);
@@ -65,10 +65,10 @@ cleanup:
     return res;
 }
 
-ssize_t syscall_routine_fstat(size_t fd, uint64_t out_stat_ptr, proc_t *proc, pml4_t *current_pml4) {
+int syscall_routine_fstat(int fd, uint64_t out_stat_ptr, proc_t *proc, pml4_t *current_pml4) {
     vfs_handle_t *handle = NULL;
 
-    ssize_t res = syscall_fs_helper_get_handle_and_check_backing_node(fd, proc, &handle);
+    int res = syscall_fs_helper_get_handle_and_check_backing_node(fd, proc, &handle);
     if (res < 0) {
         printf("syscall_routine_fstat: failed to get handle with backing node\n");
         return res;
@@ -77,7 +77,7 @@ ssize_t syscall_routine_fstat(size_t fd, uint64_t out_stat_ptr, proc_t *proc, pm
     return syscall_routine_stat_from_handle(handle, out_stat_ptr, current_pml4);
 }
 
-ssize_t syscall_routine_lstat(const char *restrict path, uint64_t out_stat_ptr, proc_t *proc, pml4_t *current_pml4) {
+int syscall_routine_lstat(const char *restrict path, uint64_t out_stat_ptr, proc_t *proc, pml4_t *current_pml4) {
     // Not implemented yet; for now, just call stat
     // TODO: once links are implemented, change this to not follow links
     return syscall_routine_stat(path, out_stat_ptr, proc, current_pml4);
